增加了舵机信号的精度;

+++ src/WritingArm.h
@@ -24,17 +24,17 @@ public:
 
     /// 设置舵机的角度, 不自动添加延迟.
     /// 任一参数给负数表示选用当前舵机的角度.
-    void setDegree(int theta, int alpha, int beta);
+    void setDegree(double theta, double alpha, double beta);
 
     /// 获取转动舵机到指定角度所需要的合适的延迟, setDegree 后添加合适延迟可以防止舵机过分抖动.
     /// 单位为毫秒
     ///
     /// 任一参数给负数表示选用当前舵机的角度.
-    double getAppropriateDelay(int theta, int alpha, int beta) const;
+    double getAppropriateDelay(double theta, double alpha, double beta) const;
 
     /// 获取舵机的角度, 0 表示 A 轴舵机, 1 表示 B 轴舵机, 2 表示 C 轴舵机.
     /// 返回 -1 表示参数不合理.
-    int getDegree(int servo) const;
+    double getDegree(int servo) const;
 
     /// 笔移动到纸面坐标 xy , 以及 z 坐标, 是对 setDegree 的封装.
     void moveTo(double x, double y, double z);
